Changed grammar. Minor change in forward models.
Body part is not assumed to automatically be a part of object, it is
also added as a terminal to grammar and sampler is required to learn
the body part as well. This makes the problem of having random parts
at the origin that are largely hidden in body part go away. Now the
sampler is able to learn true representations.